add post moment method and update permissions

+++ life_moments_app/views.py
@@ -220,48 +220,40 @@ class MomentViewSet(viewsets.ModelViewSet):
     permission_classes = [IsAuth]
 
     def create(self, request):
-        if 'profile_picture' in request.FILES:
-            file = request.FILES['profile_picture']
-            minio_client = MinioClientSingleton()
-            client = minio_client.client
-            bucket_name = minio_client.bucket_name
-            file_name = generate_unique_file_name(file.name)
-            file_path = "http://localhost:9000/life-moments/" + file_name
-
         try:
-            client.put_object(bucket_name, file_name, file, length=file.size, content_type=file.content_type)
-            print("Файл успешно загружен в Minio.")
-            request.data['profile_picture'] = file_path
-            serializer = self.serializer_class(data=request.data)
-
-            if serializer.is_valid():
-
+            ssid = request.COOKIES["session_id"]
+            if session_storage.exists(ssid):
+                username = session_storage.get(ssid).decode('utf-8')
+                user = CustomUser.objects.get(username=username)
+                if 'image' in request.FILES:
+                    file = request.FILES['image']
+                    minio_client = MinioClientSingleton()
+                    client = minio_client.client
+                    bucket_name = minio_client.bucket_name
+                    file_name = generate_unique_file_name(file.name)
+                    file_path = "http://localhost:9000/life-moments/" + file_name
+                    try:
+                        client.put_object(bucket_name, file_name, file, length=file.size, content_type=file.content_type)
+                    except Exception as e:
+                        return HttpResponseServerError('An error occurred during file upload.')
+                    
                 currentDate = datetime.now().strftime('%Y-%m-%d')
-                self.model_class.objects.create_user(
-                    username=serializer.data['username'],
-                    email=serializer.data['email'],
-                    password=serializer.data['password'],
-                    profile_picture=file_path,
-                    rating=0, # TODO: Сделать функцию расчета рейтинга
-                    registration_date=currentDate,
-                )
-                
-                random_key = str(uuid.uuid4())
-                session_storage.set(random_key, serializer.data['username'])
-
-                user_data = {
-                    "username": request.data['username'],
-                    "email": request.data['email'],
-                    "profile_picture": file_path,
-                    "rating": 0, # TODO: Сделать функцию расчета рейтинга
-                    "registration_date": currentDate
+                moment_data = {
+                    'title': request.data.get('title'),
+                    'description': request.data.get('description'),
+                    'image': file_path,
+                    'publication_date': currentDate,
+                    'id_author': user.id,
                 }
-                
-                response = Response(user_data, status=status.HTTP_201_CREATED)
-                response.set_cookie("session_id", random_key)
-                return response
-        except Exception as e:
-            print("Ошибка при загрузке файла в Minio:", str(e))
-            return HttpResponseServerError('An error occurred during file upload.')
 
-        return Response({'status': 'Error', 'error': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+                serializer = MomentSerializer(data=moment_data)
+
+                if serializer.is_valid():
+                    moment = serializer.save()
+                    return Response({'status': 'success', 'moment_id': moment.id}, status=status.HTTP_201_CREATED)
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+            return Response({'status': 'success'}, status=status.HTTP_200_OK)
+
+        except:
+            return Response({'status': 'Error'}, status=status.HTTP_400_BAD_REQUEST)
